requirements.txt
requests
pysqlite3-binary
beautifulsoup4
google
sentence-transformers
python-dotenv
faiss-cpu
numpy
pytz
langchain
langchain-google-genai
langchain-openai
langchain-chroma
langchain-community
langchain-huggingface
streamlit
schedule